style: move pending input cards into the composer slot
When an agent requests user input (questions, secrets, files, MCP auth, etc.),
the card now occupies the bottom-of-chat composer slot instead of rendering
inline in the message stream — so the thing the agent is waiting on is always
visible. The inline tool call still serves as the transcript marker.
